Change: move api_types_test.go to controlers directory for API clientset reference
