+++ Makefile
@@ -27,15 +27,11 @@ all: test manager addonctl
 
 # Run tests
 .PHONY: test
-test: test.controllers test.apis test.pkg test.cmd
+test: test.controllers test.pkg test.cmd
 
 .PHONY: test.controllers
 test.controllers:
 	go test -v -race ./controllers/... -coverprofile cover.out
-
-.PHONY: test.apis
-test.controllers:
-	go test -v -race ./apis/addon/... -coverprofile cover.out
 
 .PHONY: test.pkg
 test.pkg:
